NotNot simplification added
Copy method added as well.  A call to the simplification method returns a new
expresion that is a simplified version of the expresion the method is called on.
I can explain the rationale for this if needed.

Accompanying test for both the copy method and the simplification method have been
added as well.  I created a new test file for simplification tests it can
be merged with the main logicTest file later.