Bug in results().
An assert was added to each "conversion" method to check that the results
of each of the expresions remained the same.  This was done after finding
some inconsistent behevior.  There are test that now fail.  I did not
intruduce a bug, I have only detected the presence of one and have not been
able to locate it as of yet.  Basically don't hate me that the tests fail now.